display custom placeholders for thumbnails of files

+++ inc/display_file.php
@@ -55,37 +55,35 @@ if ($count > 0) {
             echo "<img src='img/small-trash.png' title='delete' alt='delete' /></a>";
         } // end if it is in edit mode
 
-        // THUMBNAIL GENERATION
-        // check first for the GD extension
-        if (extension_loaded('gd')) {
-            // Get file extension to display thumbnail if it's a valid extension
-            $ext = get_ext($uploads_data['real_name']);
-            if ($ext === 'jpg' ||
-                $ext === 'jpeg' ||
-                $ext === 'JPG' ||
-                $ext === 'JPEG' ||
-                $ext === 'png' ||
-                $ext === 'gif') {
+        // get file extension
+        $ext = filter_var(get_ext($uploads_data['real_name']), FILTER_SANITIZE_STRING);
+        $filepath = 'uploads/'.$uploads_data['long_name'];
+        $filesize = filesize('uploads/'.$uploads_data['long_name']);
+        $thumbpath = 'uploads/'.$uploads_data['long_name'].'_th.'.$ext;
 
-                $filepath = 'uploads/'.$uploads_data['long_name'];
-                $filesize = filesize('uploads/'.$uploads_data['long_name']);
-                $thumbpath = 'uploads/'.$uploads_data['long_name'].'_th.'.$ext;
-                // Make thumbnail only if it isn't done already and if size < 2 Mbytes
-                if (!file_exists($thumbpath) && $filesize <= 2000000) {
-                    make_thumb($filepath, $ext, $thumbpath, 150);
-                }
-                // only display the thumbnail if the file is here
-                if (file_exists($thumbpath)) {
-                    // we add rel='gallery' to the images for fancybox to display it as an album (possibility to go next/previous)
-                    echo "<a href='uploads/".$uploads_data['long_name']."' class='fancybox' rel='gallery' ";
-                    if ($uploads_data['comment'] != 'Click to add a comment') {
-                        echo "title='".$uploads_data['comment']."'";
-                    }
-                    echo "><img class='thumb' src='".$thumbpath."' alt='thumbnail' /></a>";
-                }
-            } else {
-                echo "<img class='thumb' src='img/thumb.png' alt='thumbnail' />";
+        // list of image type we can deal with the GD lib
+        $image_extensions = array('jpg', 'jpeg', 'JPG', 'JPEG', 'png', 'PNG', 'gif', 'GIF');
+        // list of extensions with a corresponding img/thumb-*.png image
+        $common_extensions = array('avi', 'csv', 'doc', 'docx', 'mov', 'pdf', 'ppt', 'rar', 'xls', 'xlsx', 'zip');
+
+        // Make thumbnail only if it isn't done already and if size < 2 Mbytes and if it's an image
+        if (!file_exists($thumbpath)
+            && $filesize <= 2000000
+            && in_array($ext, $image_extensions)) {
+            make_thumb($filepath, $ext, $thumbpath, 100);
+        }
+        // only display the thumbnail if the file is here
+        if (file_exists($thumbpath) && in_array($ext, $image_extensions)) {
+            // we add rel='gallery' to the images for fancybox to display it as an album (possibility to go next/previous)
+            echo "<a href='uploads/".$uploads_data['long_name']."' class='fancybox' rel='gallery' ";
+            if ($uploads_data['comment'] != 'Click to add a comment') {
+                echo "title='".$uploads_data['comment']."'";
             }
+            echo "><img class='thumb' src='".$thumbpath."' alt='thumbnail' /></a>";
+        } elseif (in_array($ext, $common_extensions)) {
+            echo "<img class='thumb' src='img/thumb-$ext.png' alt='' />";
+        } else { // uncommon extension without a nice image to display
+            echo "<img class='thumb' src='img/thumb.png' alt='' />";
         }
 
         // now display the name + comment with icons
